refactor(admin): extract playlists/media-players/game-settings setup-sections into ES modules (#1279 step 4b/6)
The 4 setup-sections deferred in step 4 (waiting on step-5 state
centralization) now lift out of admin.js into their own ES modules under
www/js/admin/sections/. They read/write the shared adminState object
directly — no DI closures.

- admin/constants.js — shared STORAGE keys, PLATFORM_LABELS, TAG_CATEGORIES
- admin/sections/playlists.js — list render + selection + tag-filter +
  shared selection-summary/start-button validation helpers
- admin/sections/media-players.js — speaker list + radio-selection +
  platform-capability gate (updateProviderOptions/Warning)
- admin/sections/game-settings.js — chip/toggle wiring incl. the
  music-service provider chips, localStorage load/save, summary badge,
  Title&Artist UI sync

music-service has no standalone module: its live UI is the provider chips
(game-settings) + the capability gate (media-players); only dead
setupProviderSelector code would have been a separate module.

window shims preserved: clearPlaylistFilters (inline onclick in generated
HTML), loadStatus + loadSavedSettings (cross-file consumers). admin.js
4010 -> 2965 lines. 227 tests green, build:check drift-free.
