+++ custom_components/beatify/www/js/admin.js
@@ -18,6 +18,10 @@
 // wake-lock, `_homeStartBtnHTML`, `rematchInProgress`) stay as admin.js `let`s.
 import { adminState } from './admin/state.js';
 
+// #1279 step 4b: shared constants (localStorage keys) extracted so both this
+// core and the setup-section modules import the same literals.
+import { STORAGE_LAST_PLAYER, STORAGE_GAME_SETTINGS } from './admin/constants.js';
+
 import {
     setCurrentGameResolver,
     _getAdminToken,
@@ -46,11 +50,8 @@ import {
 
 // #1279 Schritt 4/6: pure spectator-view render helpers. These are fully
 // data-in / DOM-out (no admin-private state, no cross-refs) so they lift out
-// cleanly into a unit-testable module. The four setup-sections themselves
-// (media-players, music-service, playlists, game-settings) share heavily-mutated
-// setup state with the start-game/lobby core that stays here and are deferred to
-// a post-step-5 follow-up — see the PR body. No window shims needed: every
-// caller is inside admin.js (verified) and none of these were ever global.
+// cleanly into a unit-testable module. No window shims needed: every caller is
+// inside admin.js (verified) and none of these were ever global.
 import {
     renderAdminSubmissionDots,
     renderAdminLeaderboard,
@@ -58,6 +59,46 @@ import {
     renderAdminChallengeOptions,
     _providerDisplayName,
 } from './admin/sections/render-helpers.js';
+
+// #1279 Schritt 4b/6: the four setup-sections (playlists, media-players,
+// game-settings) now share the centralized `adminState` object (step 5), so
+// they lift out into their own ES modules and read/write state directly — no DI
+// closures. The start-game/lobby core (startGame, loadStatus, BeatifyHome) stays
+// here and imports the section functions it drives. Circular-import-safe: every
+// cross-module call is event/runtime-driven, none run at module init.
+//
+// playlists.js: list render + selection + tag-filter, plus the shared
+// selection-summary / start-button-validation helpers. The intra-section
+// callees (handlePlaylistToggle, filter-bar render, etc.) are wired up inside
+// the module; admin.js core only drives the entry points below.
+// `clearPlaylistFilters` is shimmed onto `window` (below) for the inline
+// `onclick=` in the HTML the module generates.
+import {
+    renderPlaylists,
+    clearPlaylistFilters,
+    updateStartButtonState,
+} from './admin/sections/playlists.js';
+
+// media-players.js: speaker list render + radio-selection + platform-capability
+// gate (updateProviderOptions toggles the music-service provider chips). No
+// window shim: the no-players empty state's inline onclick="loadStatus()" resolves
+// to the loadStatus core fn already shimmed onto window below. admin.js core
+// drives renderMediaPlayers (from loadStatus) + handleMediaPlayerSelect (from
+// BeatifyHome.hydrateFromStorage); the rest are intra-section.
+import {
+    renderMediaPlayers,
+    handleMediaPlayerSelect,
+} from './admin/sections/media-players.js';
+
+// game-settings.js: chip/toggle wiring (language, timer, difficulty, bonus
+// flags + the music-service provider chips), the localStorage load/save
+// round-trip, the summary badge, and the Title&Artist-mode UI sync. admin.js
+// core calls setupGameSettings() + loadSavedSettings() at init; the rest are
+// intra-section. `loadSavedSettings` is shimmed onto window below for wizard.js.
+import {
+    setupGameSettings,
+    loadSavedSettings,
+} from './admin/sections/game-settings.js';
 
 // Token helpers in util.js need the live `currentGame`. The resolver reads it
 // off the shared `adminState` object (#1279 step 5), so it stays in sync across
@@ -75,6 +116,10 @@ window.buildRequestRowHtml = buildRequestRowHtml;
 window._getAdminToken = _getAdminToken;
 window._setAdminToken = _setAdminToken;
 window._adminHeaders = _adminHeaders;
+// #1279 step 4b: playlists.js generates HTML with inline onclick="clearPlaylistFilters()"
+// (empty-filter "Clear Filters" button + active-filter "Clear" link), so the
+// function must stay reachable as a window global.
+window.clearPlaylistFilters = clearPlaylistFilters;
 
 // Screen Wake Lock (#622, #1122)
 // Layer 1: navigator.wakeLock — Safari ≥16.4, Chrome, Edge, Firefox.
@@ -197,20 +242,13 @@ initAdminApi({
 let revealAdvanceInterval = null;
 let revealAdvanceOrigIcon = null;
 
-// LocalStorage keys
-const STORAGE_LAST_PLAYER = 'beatify_last_player';
-const STORAGE_GAME_SETTINGS = 'beatify_game_settings';
+// LocalStorage keys + PLATFORM_LABELS now live in ./admin/constants.js (#1279
+// step 4b) so the setup-section modules and this core share the same literals.
+// STORAGE keys are still read here (BeatifyHome hydrate, force-reset cleanup);
+// PLATFORM_LABELS moved entirely into media-players.js with its renderer.
 
 // Setup sections to hide/show as a group
 const setupSections = ['media-players', 'music-service', 'playlists', 'game-settings', 'admin-actions', 'my-requests', 'party-lights', 'tts-settings', 'ha-entities'];
-
-// Platform display labels for speaker grouping
-const PLATFORM_LABELS = {
-    music_assistant: { icon: '🎵', label: 'Music Assistant', recommended: true },
-    sonos: { icon: '🔊', label: 'Sonos' },
-    alexa_media: { icon: '📢', label: 'Alexa' },
-    alexa: { icon: '📢', label: 'Alexa' },
-};
 
 // Alias BeatifyUtils for convenience
 const utils = window.BeatifyUtils || {};
@@ -786,1101 +824,18 @@ function setupCollapsibleSections() {
     });
 }
 
-/**
- * Setup game settings controls (chips for language, timer, difficulty, toggle for artist challenge)
- */
-function setupGameSettings() {
-    // Language chips
-    document.querySelectorAll('.chip[data-lang]').forEach(chip => {
-        chip.addEventListener('click', async function() {
-            const lang = this.dataset.lang;
-            document.querySelectorAll('.chip[data-lang]').forEach(c => c.classList.remove('chip--active'));
-            this.classList.add('chip--active');
-            adminState.selectedLanguage = lang;
-            if (window.BeatifyI18n) {
-                await BeatifyI18n.setLanguage(lang);
-                BeatifyI18n.initPageTranslations();
-            }
-            updateGameSettingsSummary();
-            saveGameSettings();
-        });
-    });
-
-    // Timer chips
-    document.querySelectorAll('.chip[data-duration]').forEach(chip => {
-        chip.addEventListener('click', function() {
-            const duration = parseInt(this.dataset.duration, 10);
-            document.querySelectorAll('.chip[data-duration]').forEach(c => c.classList.remove('chip--active'));
-            this.classList.add('chip--active');
-            adminState.selectedDuration = duration;
-            updateGameSettingsSummary();
-            saveGameSettings();
-        });
-    });
-
-    // Reveal auto-advance chips (#1012)
-    document.querySelectorAll('.chip[data-reveal-advance]').forEach(chip => {
-        chip.addEventListener('click', function() {
-            adminState.revealAutoAdvance = parseInt(this.dataset.revealAdvance, 10) || 0;
-            document.querySelectorAll('.chip[data-reveal-advance]').forEach(c => c.classList.remove('chip--active'));
-            this.classList.add('chip--active');
-            saveGameSettings();
-        });
-    });
-
-    // Difficulty chips
-    document.querySelectorAll('.chip[data-difficulty]').forEach(chip => {
-        chip.addEventListener('click', function() {
-            const difficulty = this.dataset.difficulty;
-            document.querySelectorAll('.chip[data-difficulty]').forEach(c => c.classList.remove('chip--active'));
-            this.classList.add('chip--active');
-            adminState.selectedDifficulty = difficulty;
-            updateGameSettingsSummary();
-            saveGameSettings();
-        });
-    });
-
-    // Artist Challenge toggle
-    document.getElementById('artist-challenge-toggle')?.addEventListener('change', function() {
-        adminState.artistChallengeEnabled = this.checked;
-        updateGameSettingsSummary();
-        saveGameSettings();
-    });
-
-    // Movie Quiz Bonus toggle (#947)
-    document.getElementById('movie-quiz-toggle')?.addEventListener('change', function() {
-        adminState.movieQuizEnabled = this.checked;
-        updateGameSettingsSummary();
-        saveGameSettings();
-    });
-
-    // Intro Mode toggle (Issue #23)
-    document.getElementById('intro-mode-toggle')?.addEventListener('change', function() {
-        adminState.introModeEnabled = this.checked;
-        updateGameSettingsSummary();
-        saveGameSettings();
-    });
-
-    // Closest Wins toggle (Issue #442)
-    document.getElementById('closest-wins-toggle')?.addEventListener('change', function() {
-        adminState.closestWinsModeEnabled = this.checked;
-        updateGameSettingsSummary();
-        saveGameSettings();
-    });
-
-    // Title & Artist Mode toggle (#1180)
-    document.getElementById('title-artist-mode-toggle')?.addEventListener('change', function() {
-        adminState.titleArtistModeEnabled = this.checked;
-        syncTitleArtistModeUI();
-        updateGameSettingsSummary();
-        saveGameSettings();
-    });
-
-    // Provider chips (Music Service)
-    document.querySelectorAll('.chip[data-provider]').forEach(chip => {
-        chip.addEventListener('click', function() {
-            // Don't allow clicking disabled chips
-            if (this.disabled || this.classList.contains('chip--disabled')) {
-                return;
-            }
-            const provider = this.dataset.provider;
-            document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-            this.classList.add('chip--active');
-            adminState.selectedProvider = provider;
-            updateGameSettingsSummary();
-            saveGameSettings();
-            // Re-render playlists to show coverage for selected provider (preserve valid selections)
-            if (adminState.playlistData.length > 0) {
-                renderPlaylists(adminState.playlistData, '', true);
-            }
-        });
-    });
-}
-
-/**
- * Load saved settings from localStorage
- */
-async function loadSavedSettings() {
-    try {
-        const saved = localStorage.getItem(STORAGE_GAME_SETTINGS);
-        if (saved) {
-            const settings = JSON.parse(saved);
-
-            // Apply language
-            if (settings.language) {
-                adminState.selectedLanguage = settings.language;
-                document.querySelectorAll('.chip[data-lang]').forEach(c => {
-                    c.classList.toggle('chip--active', c.dataset.lang === settings.language);
-                });
-                if (window.BeatifyI18n) {
-                    await BeatifyI18n.setLanguage(settings.language);
-                    BeatifyI18n.initPageTranslations();
-                }
-            }
-
-            // Apply timer
-            if (settings.duration) {
-                adminState.selectedDuration = settings.duration;
-                document.querySelectorAll('.chip[data-duration]').forEach(c => {
-                    c.classList.toggle('chip--active', parseInt(c.dataset.duration, 10) === settings.duration);
-                });
-            }
-
-            // Apply reveal auto-advance (#1012)
-            if (typeof settings.revealAutoAdvance === 'number') {
-                adminState.revealAutoAdvance = settings.revealAutoAdvance;
-                document.querySelectorAll('.chip[data-reveal-advance]').forEach(c => {
-                    c.classList.toggle('chip--active', parseInt(c.dataset.revealAdvance, 10) === settings.revealAutoAdvance);
-                });
-            }
-
-            // Apply difficulty
-            if (settings.difficulty) {
-                adminState.selectedDifficulty = settings.difficulty;
-                document.querySelectorAll('.chip[data-difficulty]').forEach(c => {
-                    c.classList.toggle('chip--active', c.dataset.difficulty === settings.difficulty);
-                });
-            }
-
-            // Apply artist challenge
-            if (typeof settings.artistChallenge === 'boolean') {
-                adminState.artistChallengeEnabled = settings.artistChallenge;
-                const toggle = document.getElementById('artist-challenge-toggle');
-                if (toggle) toggle.checked = settings.artistChallenge;
-            }
-
-            // Apply movie quiz bonus (#947)
-            if (typeof settings.movieQuiz === 'boolean') {
-                adminState.movieQuizEnabled = settings.movieQuiz;
-                const toggle = document.getElementById('movie-quiz-toggle');
-                if (toggle) toggle.checked = settings.movieQuiz;
-            }
-
-            // Apply intro mode (Issue #23)
-            if (typeof settings.introMode === 'boolean') {
-                adminState.introModeEnabled = settings.introMode;
-                const introToggle = document.getElementById('intro-mode-toggle');
-                if (introToggle) introToggle.checked = settings.introMode;
-            }
-
-            // Apply closest wins mode (Issue #442)
-            if (typeof settings.closestWinsMode === 'boolean') {
-                adminState.closestWinsModeEnabled = settings.closestWinsMode;
-                const closestToggle = document.getElementById('closest-wins-toggle');
-                if (closestToggle) closestToggle.checked = settings.closestWinsMode;
-            }
-
-            // Apply Title & Artist mode (#1180)
-            if (typeof settings.titleArtistMode === 'boolean') {
-                adminState.titleArtistModeEnabled = settings.titleArtistMode;
-                const taToggle = document.getElementById('title-artist-mode-toggle');
-                if (taToggle) taToggle.checked = settings.titleArtistMode;
-            }
-            syncTitleArtistModeUI();
-
-            // Apply provider
-            if (settings.provider) {
-                adminState.selectedProvider = settings.provider;
-                document.querySelectorAll('.chip[data-provider]').forEach(c => {
-                    c.classList.toggle('chip--active', c.dataset.provider === settings.provider);
-                });
-            }
-        }
-    } catch (e) {
-        console.warn('Failed to load saved settings:', e);
-    }
-    // Always update summary (uses current state values)
-    updateGameSettingsSummary();
-}
-
-/**
- * Save game settings to localStorage
- */
-function saveGameSettings() {
-    try {
-        const settings = {
-            language: adminState.selectedLanguage,
-            duration: adminState.selectedDuration,
-            revealAutoAdvance: adminState.revealAutoAdvance,  // #1012
-            difficulty: adminState.selectedDifficulty,
-            artistChallenge: adminState.artistChallengeEnabled,
-            movieQuiz: adminState.movieQuizEnabled,  // #947
-            introMode: adminState.introModeEnabled,  // Issue #23
-            closestWinsMode: adminState.closestWinsModeEnabled,  // Issue #442
-            titleArtistMode: adminState.titleArtistModeEnabled,  // #1180
-            provider: adminState.selectedProvider
-        };
-        localStorage.setItem(STORAGE_GAME_SETTINGS, JSON.stringify(settings));
-    } catch (e) {
-        console.warn('Failed to save settings:', e);
-    }
-}
-
-/**
- * Update the game settings summary badge
- */
-function updateGameSettingsSummary() {
-    const summary = document.getElementById('game-settings-summary');
-    if (!summary) return;
-
-    const difficultyLabels = { easy: 'Easy', normal: 'Normal', hard: 'Hard' };
-    const langLabels = { en: 'EN', de: 'DE', es: 'ES' };
-    // #1180: year-round bonuses are suppressed while TA mode is on, so the badge
-    // hides their icons too — but the underlying flags stay the host's untouched
-    // source of truth (so toggling TA off restores them).
-    const yearRoundActive = !adminState.titleArtistModeEnabled;
-    const artistIcon = (yearRoundActive && adminState.artistChallengeEnabled) ? ' • 🎤' : '';
-    const movieIcon = (yearRoundActive && adminState.movieQuizEnabled) ? ' • 🎬' : '';  // #947
-    const introIcon = (yearRoundActive && adminState.introModeEnabled) ? ' • ⚡' : '';  // Issue #23
-    const closestIcon = (yearRoundActive && adminState.closestWinsModeEnabled) ? ' • 🎯' : '';  // Issue #442
-    const taIcon = adminState.titleArtistModeEnabled ? ' • 🎵' : '';  // #1180
-
-    summary.textContent = `${difficultyLabels[adminState.selectedDifficulty] || 'Normal'} • ${adminState.selectedDuration}s • ${langLabels[adminState.selectedLanguage] || 'EN'}${taIcon}${artistIcon}${movieIcon}${introIcon}${closestIcon}`;
-}
-
-/**
- * #1180: Title & Artist mode replaces the year round, so the year-only
- * bonuses (artist challenge, movie quiz, intro, closest wins) have nothing to
- * attach to. Hide and disable their setting-groups while TA mode is on.
- *
- * This is purely a visibility/disabled-state sync — it does NOT mutate the
- * year-round flags or the checkboxes. The host's real bonus preferences stay
- * the single source of truth (in the in-memory flags, the checkboxes, and
- * localStorage), so the save → reload → toggle-off cycle is lossless. The
- * actual suppression (forcing year-round bonuses off when TA mode is on) is
- * applied only when building the start-game payload, in startGame(), via
- * applyTitleArtistBonusPrecedence(). Forcing the flags off here instead would
- * persist false to localStorage and silently destroy the host's choices on
- * the next reload.
- */
-function syncTitleArtistModeUI() {
-    // #1180: only the truly-incompatible modes are hidden in TA mode. Movie
-    // quiz and intro mode are compatible bonuses, so they stay available.
-    var ids = ['artist-challenge-toggle', 'closest-wins-toggle'];
-    ids.forEach(function(id) {
-        var input = document.getElementById(id);
-        if (!input) return;
-        var group = input.closest('.setting-group');
-        if (group) group.classList.toggle('hidden', adminState.titleArtistModeEnabled);
-        input.disabled = adminState.titleArtistModeEnabled;
-    });
-    // #1180 polish: year-distance difficulty doesn't apply in TA mode. Hide the
-    // chips + year hint and show the fixed T&I scoring summary in their place.
-    var diffRow = document.getElementById('admin-difficulty-row');
-    if (diffRow) diffRow.classList.toggle('hidden', adminState.titleArtistModeEnabled);
-    var diffHint = document.getElementById('admin-difficulty-hint');
-    if (diffHint) diffHint.classList.toggle('hidden', adminState.titleArtistModeEnabled);
-    var taSummary = document.getElementById('admin-difficulty-ta-summary');
-    if (taSummary) taSummary.classList.toggle('hidden', !adminState.titleArtistModeEnabled);
-}
-
-/**
- * Update media player summary badge
- * @param {string} playerName - Friendly name of selected player
- */
-function updateMediaPlayerSummary(playerName) {
-    const summary = document.getElementById('media-player-summary');
-    if (summary) {
-        summary.textContent = playerName || 'Select...';
-    }
-}
-
-// groupPlayersByPlatform moved to ./admin/util.js (#1279 step 2)
-
-/**
- * Render media players list grouped by platform with capability info
- * Filters out unavailable players
- * @param {Array} players
- */
-function renderMediaPlayers(players) {
-    const container = document.getElementById('media-players-list');
-    // Remove data-i18n and skeleton state when real content renders
-    container?.removeAttribute('data-i18n');
-    container?.removeAttribute('aria-busy');
-    container?.classList.remove('skeleton-list');
-    const totalPlayers = players ? players.length : 0;
-
-    // Reset selection state
-    adminState.selectedMediaPlayer = null;
-
-    // Filter out unavailable players
-    const availablePlayers = (players || []).filter(p => p.state !== 'unavailable');
-
-    // Hide validation message when showing empty states (avoid redundant messaging)
-    const validationMsg = document.getElementById('media-player-validation-msg');
-
-    if (totalPlayers === 0) {
-        // No compatible players found - show setup message with MA link
-        container.innerHTML = `
-            <div class="no-players-message">
-                <h3>🎵 No Compatible Players Found</h3>
-                <p>Beatify works with Music Assistant, Sonos, and Alexa players.</p>
-                <p><strong>Recommended:</strong> Install Music Assistant for the best experience with any speaker.</p>
-                <div class="button-group">
-                    <a href="https://music-assistant.io/getting-started/"
-                       target="_blank" class="btn btn-secondary">
-                        📖 Music Assistant Setup Guide
-                    </a>
-                    <button onclick="loadStatus()" class="btn btn-primary">
-                        🔄 Refresh
-                    </button>
-                </div>
-            </div>
-        `;
-        if (validationMsg) {
-            validationMsg.classList.add('hidden');
-        }
-        // Disable start button when no players
-        const startBtn = document.getElementById('start-game');
-        if (startBtn) startBtn.disabled = true;
-        return;
-    }
-
-    if (availablePlayers.length === 0) {
-        // Players exist but all unavailable
-        const docsLink = adminState.mediaPlayerDocsUrl
-            ? `<a href="${utils.escapeHtml(adminState.mediaPlayerDocsUrl)}" target="_blank" rel="noopener">Troubleshooting</a>`
-            : '';
-        container.innerHTML = `
-            <div class="empty-state">
-                <p class="status-error">All media players are unavailable. Check your devices are powered on.</p>
-                ${docsLink ? `<p style="margin-top: 12px;">${docsLink}</p>` : ''}
-            </div>
-        `;
-        if (validationMsg) {
-            validationMsg.classList.add('hidden');
-        }
-        return;
-    }
-
-    // Render all players with platform badges on each item
-    container.innerHTML = availablePlayers.map(player => renderPlayerItem(player)).join('');
-    attachPlayerSelectionHandlers();
-
-    // Try to auto-select last used player from localStorage
-    const lastPlayerId = localStorage.getItem(STORAGE_LAST_PLAYER);
-    if (lastPlayerId) {
-        const lastPlayerRadio = container.querySelector(`[data-entity-id="${lastPlayerId}"]`);
-        if (lastPlayerRadio) {
-            lastPlayerRadio.checked = true;
-            handleMediaPlayerSelect(lastPlayerRadio, true); // true = skip localStorage save
-            // Collapse section since we have a valid selection
-            const section = document.getElementById('media-players');
-            if (section) {
-                section.classList.add('collapsed');
-                const toggle = document.getElementById('media-players-toggle');
-                if (toggle) toggle.setAttribute('aria-expanded', 'false');
-            }
-        }
-    }
-}
-
-/**
- * Render a single player item with platform badge and capability data attributes
- * @param {Object} player - Player object from backend
- * @returns {string} HTML string
- */
-function renderPlayerItem(player) {
-    const info = PLATFORM_LABELS[player.platform] || { icon: '🔈', label: player.platform };
-    const platformBadge = `<span class="platform-badge platform-badge--${utils.escapeHtml(player.platform)}">${info.icon} ${info.label}</span>`;
-
-    return `
-        <div class="media-player-item list-item is-selectable"
-             data-entity-id="${utils.escapeHtml(player.entity_id)}"
-             data-platform="${utils.escapeHtml(player.platform)}"
-             data-supports-spotify="${player.supports_spotify}"
-             data-supports-apple-music="${player.supports_apple_music}"
-             data-supports-youtube-music="${player.supports_youtube_music}"
-             data-supports-tidal="${player.supports_tidal}"
-             data-supports-deezer="${player.supports_deezer}"
-             data-supports-amazon-music="${player.supports_amazon_music}">
-            <label class="radio-label">
-                <input type="radio"
-                       class="media-player-radio"
-                       name="media-player"
-                       data-entity-id="${utils.escapeHtml(player.entity_id)}"
-                       data-state="${utils.escapeHtml(player.state)}"
-                       data-platform="${utils.escapeHtml(player.platform)}"
-                       data-supports-spotify="${player.supports_spotify}"
-                       data-supports-apple-music="${player.supports_apple_music}"
-                       data-supports-youtube-music="${player.supports_youtube_music}"
-                       data-supports-tidal="${player.supports_tidal}"
-                       data-supports-deezer="${player.supports_deezer}"
-                       data-supports-amazon-music="${player.supports_amazon_music}">
-                <span class="player-info">
-                    <span class="player-name">${utils.escapeHtml(player.friendly_name)}</span>
-                    ${platformBadge}
-                </span>
-            </label>
-            <span class="meta">
-                <span class="state-dot state-${utils.escapeHtml(player.state)}"></span>
-                ${utils.escapeHtml(player.state)}
-            </span>
-        </div>
-    `;
-}
-
-/**
- * Attach event handlers to player selection elements
- */
-function attachPlayerSelectionHandlers() {
-    const container = document.getElementById('media-players-list');
-    if (!container) return;
-
-    // Attach event listeners to radio buttons
-    container.querySelectorAll('.media-player-radio').forEach(radio => {
-        radio.addEventListener('change', function() {
-            handleMediaPlayerSelect(this);
-        });
-    });
-
-    // Make entire row clickable (for hidden input UX)
-    container.querySelectorAll('.media-player-item').forEach(item => {
-        item.addEventListener('click', function(e) {
-            // Don't double-trigger if clicking on the radio or within the label
-            if (e.target.classList.contains('media-player-radio') || e.target.closest('.radio-label')) return;
-            const radio = item.querySelector('.media-player-radio');
-            if (radio && !radio.checked) {
-                radio.checked = true;
-                handleMediaPlayerSelect(radio);
-            }
-        });
-    });
-}
-
-/**
- * Handle media player radio button selection (AC4)
- * Updates provider options based on platform capabilities.
- * @param {HTMLInputElement} radio
- * @param {boolean} skipSave - If true, don't save to localStorage (used for auto-select)
- */
-function handleMediaPlayerSelect(radio, skipSave = false) {
-    const entityId = radio.dataset.entityId;
-    const state = radio.dataset.state;
-    const platform = radio.dataset.platform;
-    const supportsSpotify = radio.dataset.supportsSpotify === 'true';
-    const supportsAppleMusic = radio.dataset.supportsAppleMusic === 'true';
-    const supportsYoutubeMusic = radio.dataset.supportsYoutubeMusic === 'true';
-    const supportsTidal = radio.dataset.supportsTidal === 'true';
-    const supportsDeezer = radio.dataset.supportsDeezer === 'true';
-    const supportsAmazonMusic = radio.dataset.supportsAmazonMusic === 'true';
-
-    // Update module state with platform capabilities
-    adminState.selectedMediaPlayer = {
-        entityId,
-        state,
-        platform,
-        supportsSpotify,
-        supportsAppleMusic,
-        supportsYoutubeMusic,
-        supportsTidal,
-        supportsDeezer,
-        supportsAmazonMusic,
-    };
-
-    // Update visual selection
-    document.querySelectorAll('.media-player-item').forEach(item => {
-        item.classList.remove('is-selected');
-    });
-    const playerItem = radio.closest('.media-player-item');
-    playerItem.classList.add('is-selected');
-
-    // Get player name for summary
-    const playerName = playerItem.querySelector('.player-name')?.textContent?.trim() || entityId;
-    updateMediaPlayerSummary(playerName);
-
-    // Show Music Service section
-    const musicServiceSection = document.getElementById('music-service');
-    if (musicServiceSection) {
-        musicServiceSection.classList.remove('hidden');
-    }
-
-    // Update provider options based on platform capabilities
-    updateProviderOptions(adminState.selectedMediaPlayer);
-
-    // Update warning message
-    updateProviderWarning(adminState.selectedMediaPlayer);
-
-    // Save to localStorage
-    if (!skipSave) {
-        try {
-            localStorage.setItem(STORAGE_LAST_PLAYER, entityId);
-        } catch (e) {
-            console.warn('Failed to save last player:', e);
-        }
-    }
-
-    updateStartButtonState();
-}
-
-/**
- * Update provider button states based on selected player capabilities
- * @param {Object} player - Selected player with capability flags
- */
-function updateProviderOptions(player) {
-    const spotifyBtn = document.querySelector('.chip[data-provider="spotify"]');
-    const appleBtn = document.querySelector('.chip[data-provider="apple_music"]');
-    const youtubeBtn = document.querySelector('.chip[data-provider="youtube_music"]');
-    const tidalBtn = document.querySelector('.chip[data-provider="tidal"]');
-    const deezerBtn = document.querySelector('.chip[data-provider="deezer"]');
-    const amazonBtn = document.querySelector('.chip[data-provider="amazon_music"]');
-
-    if (spotifyBtn) {
-        spotifyBtn.disabled = !player.supportsSpotify;
-        spotifyBtn.classList.toggle('chip--disabled', !player.supportsSpotify);
-    }
-
-    if (appleBtn) {
-        appleBtn.disabled = !player.supportsAppleMusic;
-        appleBtn.classList.toggle('chip--disabled', !player.supportsAppleMusic);
-    }
-
-    if (youtubeBtn) {
-        youtubeBtn.disabled = !player.supportsYoutubeMusic;
-        youtubeBtn.classList.toggle('chip--disabled', !player.supportsYoutubeMusic);
-    }
-
-    if (tidalBtn) {
-        tidalBtn.disabled = !player.supportsTidal;
-        tidalBtn.classList.toggle('chip--disabled', !player.supportsTidal);
-    }
-
-    if (deezerBtn) {
-        deezerBtn.disabled = !player.supportsDeezer;
-        deezerBtn.classList.toggle('chip--disabled', !player.supportsDeezer);
-    }
-
-    if (amazonBtn) {
-        amazonBtn.disabled = !player.supportsAmazonMusic;
-        amazonBtn.classList.toggle('chip--disabled', !player.supportsAmazonMusic);
-    }
-
-    // If current selection is now disabled, switch to Spotify
-    if (adminState.selectedProvider === 'apple_music' && !player.supportsAppleMusic) {
-        // Update UI
-        document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-        if (spotifyBtn) spotifyBtn.classList.add('chip--active');
-        adminState.selectedProvider = 'spotify';
-    }
-
-    if (adminState.selectedProvider === 'youtube_music' && !player.supportsYoutubeMusic) {
-        // Update UI
-        document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-        if (spotifyBtn) spotifyBtn.classList.add('chip--active');
-        adminState.selectedProvider = 'spotify';
-    }
-
-    if (adminState.selectedProvider === 'tidal' && !player.supportsTidal) {
-        // Update UI
-        document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-        if (spotifyBtn) spotifyBtn.classList.add('chip--active');
-        adminState.selectedProvider = 'spotify';
-    }
-
-    if (adminState.selectedProvider === 'deezer' && !player.supportsDeezer) {
-        // Update UI
-        document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-        if (spotifyBtn) spotifyBtn.classList.add('chip--active');
-        adminState.selectedProvider = 'spotify';
-    }
-
-    if (adminState.selectedProvider === 'amazon_music' && !player.supportsAmazonMusic) {
-        // Update UI
-        document.querySelectorAll('.chip[data-provider]').forEach(c => c.classList.remove('chip--active'));
-        if (spotifyBtn) spotifyBtn.classList.add('chip--active');
-        adminState.selectedProvider = 'spotify';
-    }
-
-    // Show hint for disabled providers
-    const hint = document.getElementById('provider-hint');
-    if (hint) {
-        const maSpeakerNeeded = [];
-        if (!player.supportsAppleMusic) maSpeakerNeeded.push('Apple Music');
-        if (!player.supportsYoutubeMusic) maSpeakerNeeded.push('YouTube Music');
-        if (!player.supportsTidal) maSpeakerNeeded.push('Tidal');
-        if (!player.supportsDeezer) maSpeakerNeeded.push('Deezer');
-
-        const hintParts = [];
-        if (maSpeakerNeeded.length > 0) {
-            hintParts.push(`${maSpeakerNeeded.join(' and ')} require${maSpeakerNeeded.length === 1 ? 's' : ''} a Music Assistant speaker`);
-        }
-        if (!player.supportsAmazonMusic) {
-            hintParts.push('Amazon Music requires an Amazon Echo (alexa_media)');
-        }
-
-        if (hintParts.length > 0) {
-            hint.textContent = hintParts.join(' · ');
-            hint.classList.remove('hidden');
-        } else {
-            hint.classList.add('hidden');
-        }
-    }
-}
-
-/**
- * Update provider warning based on selected speaker platform
- * Shows setup requirements and caveats per platform
- * @param {Object} player - Selected player with platform info
- */
-function updateProviderWarning(player) {
-    const warningEl = document.getElementById('provider-warning');
-    if (!warningEl) return;
-
-    const platformInfo = {
-        music_assistant: {
-            warning: 'Premium account must be configured in Music Assistant',
-        },
-        sonos: {
-            warning: 'Spotify must be linked in Sonos app',
-        },
-        alexa_media: {
-            warning: 'Service must be linked in Alexa app',
-            caveat: 'Uses voice search - may occasionally play a different version of the song',
-        },
-        alexa: {
-            warning: 'Service must be linked in Alexa app',
-            caveat: 'Uses voice search - may occasionally play a different version of the song',
-        },
-    };
-
-    const info = platformInfo[player.platform];
-    if (info) {
-        let html = `<p>⚠️ ${utils.escapeHtml(info.warning)}</p>`;
-        if (info.caveat) {
-            html += `<p class="warning-caveat">ℹ️ ${utils.escapeHtml(info.caveat)}</p>`;
-        }
-        warningEl.innerHTML = html;
-        warningEl.classList.remove('hidden');
-    } else {
-        warningEl.classList.add('hidden');
-    }
-}
-
-/**
- * Render playlists list with checkboxes for valid playlists
- * @param {Array} playlists
- * @param {string} playlistDir
- * @param {boolean} preserveSelection - If true, preserve valid selections (used when provider changes)
- */
-function renderPlaylists(playlists, playlistDir, preserveSelection = false) {
-    const container = document.getElementById('playlists-list');
-    // Remove data-i18n and skeleton state when real content renders
-    container?.removeAttribute('data-i18n');
-    container?.removeAttribute('aria-busy');
-    container?.classList.remove('skeleton-list');
-
-    // Store previous selections before reset (for preserveSelection mode)
-    const previousSelections = preserveSelection ? [...adminState.selectedPlaylists] : [];
-
-    // Reset selection state
-    adminState.selectedPlaylists = [];
-    adminState.playlistData = playlists || [];
-
-    // Render filter bar (Issue #70)
-    renderPlaylistFilterBar(adminState.playlistData);
-
-    // Filter playlists based on active filters (Issue #70 - Option B)
-    // Uses AND logic: playlist must match ALL selected category filters
-    let filteredPlaylists = adminState.playlistData;
-    if (!adminState.activeFilterTags.includes('all') && adminState.activeFilterTags.length > 0) {
-        filteredPlaylists = adminState.playlistData.filter(p => {
-            const playlistTags = p.tags || [];
-            // Playlist must contain ALL active filter tags (AND logic)
-            return adminState.activeFilterTags.every(tag => playlistTags.includes(tag));
-        });
-    }
-
-    // Check if we have any valid playlists
-    const hasValidPlaylists = adminState.playlistData.some(p => p.is_valid);
-
-    if (!adminState.playlistData || adminState.playlistData.length === 0) {
-        // AC2: No playlists error with documentation link
-        const docsLink = adminState.playlistDocsUrl
-            ? `<a href="${utils.escapeHtml(adminState.playlistDocsUrl)}" target="_blank" rel="noopener">How to create playlists</a>`
-            : '';
-        container.innerHTML = `
-            <div class="empty-state">
-                <p class="status-error">No playlists found. Add playlist JSON files to:</p>
-                <p style="font-size: 14px;"><code>${utils.escapeHtml(playlistDir)}</code></p>
-                ${docsLink ? `<p style="margin-top: 12px;">${docsLink}</p>` : ''}
-            </div>
-        `;
-        // Hide start button when no playlists (Story 9.10)
-        document.getElementById('start-game')?.classList.add('hidden');
-        return;
-    }
-
-    // Show message if filter results in no playlists (Issue #70)
-    if (filteredPlaylists.length === 0) {
-        container.innerHTML = `
-            <div class="empty-state">
-                <p>No playlists match the selected filter.</p>
-                <button type="button" class="btn btn-secondary" onclick="clearPlaylistFilters()">Clear Filters</button>
-            </div>
-        `;
-        return;
-    }
-
-    container.innerHTML = filteredPlaylists.map(playlist => {
-        if (playlist.is_valid) {
-            // AC1: Valid playlists with checkbox
-            const songCount = playlist.song_count || 0;
-            const spotifyCount = playlist.spotify_count || 0;
-            const appleMusicCount = playlist.apple_music_count || 0;
-            const youtubeMusicCount = playlist.youtube_music_count || 0;
-            const tidalCount = playlist.tidal_count || 0;
-            const deezerCount = playlist.deezer_count || 0;
-            // Amazon Music uses Alexa text search — all songs are playable.
-            const amazonMusicCount = playlist.amazon_music_count || songCount;
-
-            // Get provider count based on selected provider
-            let providerCount = songCount;
-            if (adminState.selectedProvider === 'spotify') {
-                providerCount = spotifyCount || songCount; // fallback for legacy playlists
-            } else if (adminState.selectedProvider === 'apple_music') {
-                providerCount = appleMusicCount;
-            } else if (adminState.selectedProvider === 'youtube_music') {
-                providerCount = youtubeMusicCount;
-            } else if (adminState.selectedProvider === 'tidal') {
-                providerCount = tidalCount;
-            } else if (adminState.selectedProvider === 'deezer') {
-                providerCount = deezerCount;
-            } else if (adminState.selectedProvider === 'amazon_music') {
-                providerCount = amazonMusicCount;
-            }
-
-            // Disable playlist if no songs for selected provider
-            const isDisabled = providerCount === 0;
-            const disabledClass = isDisabled ? 'is-disabled' : '';
-            const disabledAttr = isDisabled ? 'disabled' : '';
-
-            // Build coverage indicator
-            let coverageHtml = '';
-            if (providerCount < songCount) {
-                const coverageClass = providerCount === 0
-                    ? 'playlist-coverage playlist-coverage--none'
-                    : 'playlist-coverage playlist-coverage--warning';
-                coverageHtml = `<span class="${coverageClass}">${providerCount}/${songCount}</span>`;
-            }
-
-            return `
-                <div class="playlist-item list-item ${isDisabled ? '' : 'is-selectable'} ${disabledClass}"
-                     data-provider-count="${providerCount}"
-                     data-tags="${utils.escapeHtml((playlist.tags || []).join(','))}">
-                    <label class="checkbox-label">
-                        <input type="checkbox"
-                               class="playlist-checkbox"
-                               data-path="${utils.escapeHtml(playlist.path)}"
-                               data-song-count="${utils.escapeHtml(String(songCount))}"
-                               data-provider-count="${providerCount}"
-                               ${disabledAttr}>
-                        <span class="playlist-name">${utils.escapeHtml(playlist.name)}</span>
-                    </label>
-                    <span class="meta">${coverageHtml || utils.escapeHtml(String(songCount))} songs</span>
-                </div>
-            `;
-        } else {
-            // Invalid playlists: no checkbox, greyed out
-            const errorMsg = (playlist.errors && playlist.errors[0]) || 'Unknown error';
-            return `
-                <div class="list-item is-invalid">
-                    <span class="name">${utils.escapeHtml(playlist.name)}</span>
-                    <span class="meta">Invalid: ${utils.escapeHtml(errorMsg)}</span>
-                </div>
-            `;
-        }
-    }).join('');
-
-    // Attach event listeners to checkboxes (instead of inline handlers)
-    container.querySelectorAll('.playlist-checkbox').forEach(checkbox => {
-        checkbox.addEventListener('change', function() {
-            handlePlaylistToggle(this);
-        });
-    });
-
-    // Make entire row clickable (for hidden input UX)
-    container.querySelectorAll('.playlist-item.is-selectable').forEach(item => {
-        item.addEventListener('click', function(e) {
-            // Don't double-trigger if clicking on the checkbox or label
-            if (e.target.classList.contains('playlist-checkbox') || e.target.closest('.checkbox-label')) return;
-            const checkbox = item.querySelector('.playlist-checkbox');
-            if (checkbox) {
-                checkbox.checked = !checkbox.checked;
-                handlePlaylistToggle(checkbox);
-            }
-        });
-    });
-
-    // Restore valid selections when preserving (provider change)
-    if (preserveSelection && previousSelections.length > 0) {
-        previousSelections.forEach(prev => {
-            const checkbox = container.querySelector(`.playlist-checkbox[data-path="${CSS.escape(prev.path)}"]`);
-            if (checkbox && !checkbox.disabled) {
-                checkbox.checked = true;
-                const providerCount = parseInt(checkbox.dataset.providerCount, 10) || 0;
-                const item = checkbox.closest('.playlist-item');
-                if (providerCount > 0) {
-                    adminState.selectedPlaylists.push({ path: prev.path, songCount: providerCount });
-                    item?.classList.add('is-selected');
-                }
-            }
-        });
-    }
-
-    // Show start button if we have valid playlists (Story 9.10)
-    if (hasValidPlaylists) {
-        document.getElementById('start-game')?.classList.remove('hidden');
-    } else {
-        document.getElementById('start-game')?.classList.add('hidden');
-    }
-
-    // Restore previously saved playlist selections from localStorage (mirrors the
-    // last-player auto-restore in renderMediaPlayers). Without this, the wizard's
-    // selections get wiped every time loadStatus() re-renders the playlist list.
-    if (adminState.selectedPlaylists.length === 0) {
-        try {
-            const raw = localStorage.getItem(STORAGE_GAME_SETTINGS);
-            const saved = raw ? JSON.parse(raw) : null;
-            const savedPaths = Array.isArray(saved?.selectedPlaylists)
-                ? saved.selectedPlaylists.map((p) => (typeof p === 'string' ? p : p.path)).filter(Boolean)
-                : [];
-            savedPaths.forEach((path) => {
-                const checkbox = container.querySelector(`.playlist-checkbox[data-path="${CSS.escape(path)}"]`);
-                if (checkbox && !checkbox.disabled) {
-                    checkbox.checked = true;
-                    const providerCount = parseInt(checkbox.dataset.providerCount, 10) || 0;
-                    if (providerCount > 0 && !adminState.selectedPlaylists.some((p) => p.path === path)) {
-                        adminState.selectedPlaylists.push({ path, songCount: providerCount });
-                        checkbox.closest('.playlist-item')?.classList.add('is-selected');
-                    }
-                }
-            });
-        } catch (e) { console.warn('[Beatify] restore saved playlists failed:', e); }
-    }
-
-    // Initialize summary as hidden
-    updateSelectionSummary();
-    updateStartButtonState();
-}
-
-/**
- * Handle playlist checkbox toggle
- * @param {HTMLInputElement} checkbox
- */
-function handlePlaylistToggle(checkbox) {
-    const path = checkbox.dataset.path;
-    // Use provider-specific count for selection tracking
-    const providerCount = parseInt(checkbox.dataset.providerCount, 10) || 0;
-    const item = checkbox.closest('.playlist-item');
-
-    if (checkbox.checked) {
-        // Prevent duplicate selections
-        if (!adminState.selectedPlaylists.some(p => p.path === path)) {
-            adminState.selectedPlaylists.push({ path, songCount: providerCount });
-        }
-        item.classList.add('is-selected');
-    } else {
-        adminState.selectedPlaylists = adminState.selectedPlaylists.filter(p => p.path !== path);
-        item.classList.remove('is-selected');
-    }
-
-    updateSelectionSummary();
-    updateStartButtonState();
-}
-
-/**
- * Render the playlist filter bar with tag buttons (Issue #70)
- * @param {Array} playlists
- */
-// Tag category definitions for dropdown filters
-const TAG_CATEGORIES = {
-    decade: {
-        label: 'Decade',
-        tags: ['1960s', '1970s', '1980s', '1990s', '2000s']
-    },
-    style: {
-        label: 'Style',
-        tags: ['rock', 'pop', 'ballads', 'electronic', 'eurodance', 'yacht-rock', 'soft-rock', 'pop-punk', 'schlager', 'party', 'britpop', 'british-invasion', 'classic-rock', 'dance', 'disco', 'funk', 'hip-hop', 'latin', 'merengue', 'motown', 'r&b', 'salsa', 'soul']
-    },
-    region: {
-        label: 'Region',
-        tags: ['international', 'german', 'dutch', 'spanish']
-    },
-    special: {
-        label: 'Special',
-        tags: ['movies', 'soundtrack', 'eurovision', 'carnival', 'classics', 'contest', 'mixed', 'one-hit', 'top-hits']
-    }
-};
-
-// Active filter state per category lives in adminState.activeFilters (step 5).
-
-function renderPlaylistFilterBar(playlists) {
-    const filterBar = document.getElementById('playlist-filter-bar');
-    if (!filterBar) return;
-
-    // Extract unique tags from all playlists
-    const availableTags = new Set();
-    playlists.forEach(p => {
-        (p.tags || []).forEach(tag => availableTags.add(tag));
-    });
-
-    // If no tags found, hide filter bar
-    if (availableTags.size === 0) {
-        filterBar.classList.add('hidden');
-        return;
-    }
-
-    // Capitalize first letter helper
-    const capitalize = (str) => str.charAt(0).toUpperCase() + str.slice(1);
-
-    // Build dropdown HTML for each category
-    let html = '<div class="filter-dropdowns">';
-    
-    Object.entries(TAG_CATEGORIES).forEach(([categoryKey, category]) => {
-        // Filter to only tags that exist in playlists
-        const categoryTags = category.tags.filter(tag => availableTags.has(tag));
-        
-        if (categoryTags.length === 0) return;
-        
-        const currentValue = adminState.activeFilters[categoryKey] || '';
-        
-        html += `
-            <select class="filter-dropdown" data-category="${categoryKey}">
-                <option value="">${category.label}</option>
-                ${categoryTags.map(tag => {
-                    const selected = currentValue === tag ? 'selected' : '';
-                    return `<option value="${utils.escapeHtml(tag)}" ${selected}>${capitalize(tag)}</option>`;
-                }).join('')}
-            </select>
-        `;
-    });
-    
-    html += '</div>';
-
-    // Show active filters summary
-    const activeFiltersList = Object.entries(adminState.activeFilters)
-        .filter(([_, value]) => value)
-        .map(([_, value]) => capitalize(value));
-    
-    if (activeFiltersList.length > 0) {
-        html += `
-            <div class="filter-summary">
-                <span class="filter-summary-text">Showing: ${activeFiltersList.join(' • ')}</span>
-                <button type="button" class="filter-clear" onclick="clearPlaylistFilters()">Clear</button>
-            </div>
-        `;
-    }
-
-    filterBar.innerHTML = html;
-    filterBar.classList.remove('hidden');
-
-    // Attach event listeners to dropdowns
-    filterBar.querySelectorAll('.filter-dropdown').forEach(select => {
-        select.addEventListener('change', function() {
-            handleFilterDropdownChange(this.dataset.category, this.value);
-        });
-    });
-}
-
-/**
- * Handle filter dropdown change (Issue #70 - Option B)
- * @param {string} category - The filter category (decade, style, region, special)
- * @param {string} value - The selected tag value
- */
-function handleFilterDropdownChange(category, value) {
-    adminState.activeFilters[category] = value;
-    
-    // Update adminState.activeFilterTags for compatibility with existing filter logic
-    updateActiveFilterTags();
-    
-    // Re-render playlists with new filter
-    renderPlaylists(adminState.playlistData, '', true);
-}
-
-/**
- * Update adminState.activeFilterTags array from adminState.activeFilters object
- */
-function updateActiveFilterTags() {
-    const selectedTags = Object.values(adminState.activeFilters).filter(v => v);
-    adminState.activeFilterTags = selectedTags.length > 0 ? selectedTags : ['all'];
-}
-
-/**
- * Clear all playlist filters (Issue #70)
- */
-function clearPlaylistFilters() {
-    adminState.activeFilters = {
-        decade: '',
-        style: '',
-        region: '',
-        special: ''
-    };
-    adminState.activeFilterTags = ['all'];
-    renderPlaylists(adminState.playlistData, '', true);
-}
-
-// Expose clearPlaylistFilters globally for onclick handler
-window.clearPlaylistFilters = clearPlaylistFilters;
-
-/**
- * Calculate total songs from selected playlists
- * @returns {number}
- */
-function calculateTotalSongs() {
-    return adminState.selectedPlaylists.reduce((sum, p) => sum + p.songCount, 0);
-}
-
-/**
- * Update the selection summary display
- */
-function updateSelectionSummary() {
-    const summary = document.getElementById('playlist-summary');
-    const selectedCount = document.getElementById('selected-count');
-    const totalSongs = document.getElementById('total-songs');
-
-    // Null check for DOM elements
-    if (!summary || !selectedCount || !totalSongs) {
-        return;
-    }
-
-    if (adminState.selectedPlaylists.length === 0) {
-        summary.classList.add('hidden');
-    } else {
-        summary.classList.remove('hidden');
-        selectedCount.textContent = adminState.selectedPlaylists.length;
-        totalSongs.textContent = calculateTotalSongs();
-    }
-}
-
-/**
- * Update start button enabled/disabled state and validation messages
- * Checks for both playlist AND media player selection
- */
-function updateStartButtonState() {
-    const btn = document.getElementById('start-game');
-    const playlistMsg = document.getElementById('playlist-validation-msg');
-    const mediaPlayerMsg = document.getElementById('media-player-validation-msg');
-
-    if (!btn) {
-        return;
-    }
-
-    const noPlaylist = adminState.selectedPlaylists.length === 0;
-    const noMediaPlayer = adminState.selectedMediaPlayer === null;
-
-    // Disable button if either selection is missing
-    btn.disabled = noPlaylist || noMediaPlayer;
-
-    // Show/hide playlist validation message
-    if (playlistMsg) {
-        playlistMsg.classList.toggle('hidden', !noPlaylist);
-    }
-
-    // Show/hide media player validation message
-    if (mediaPlayerMsg) {
-        mediaPlayerMsg.classList.toggle('hidden', !noMediaPlayer);
-    }
-}
+// Game-Settings section (setupGameSettings chip/toggle wiring incl. the
+// music-service provider chips, loadSavedSettings/saveGameSettings persistence,
+// updateGameSettingsSummary, syncTitleArtistModeUI) moved to
+// ./admin/sections/game-settings.js (#1279 step 4b). Imported at top of file.
+
+// Media-Players section (renderMediaPlayers, renderPlayerItem, selection
+// handlers, updateProviderOptions/Warning, summary) moved to
+// ./admin/sections/media-players.js (#1279 step 4b). Imported at top of file.
+
+// Playlists section (renderPlaylists, handlePlaylistToggle, filter bar,
+// selection-summary + start-button validation) moved to
+// ./admin/sections/playlists.js (#1279 step 4b). Imported at top of file.
 
 // escapeHtml moved to BeatifyUtils
